% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{adv_age}
\alias{adv_age}
\title{Advance age of all individuals in the local community}
\usage{
adv_age(lc)
}
\arguments{
\item{lc}{An object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}
}
\value{
A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).
}
\description{
Advance age of all individuals in the local community
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh)
lc = recruit(lc = lc, mc = mc, lh = lh)
adv_age(lc)

}
