% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{braycurtis}
\alias{braycurtis}
\title{Helper, Bray-Curtis similarity}
\usage{
braycurtis(com1, com2)
}
\arguments{
\item{com1}{Tibble for community 1 with two columns: species and n}

\item{com2}{Tibble for community 2 with two columns: species and n}
}
\value{
Bray-Curtis index
}
\description{
Helper, Bray-Curtis similarity
}
