% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{crun_sim_}
\alias{crun_sim_}
\title{Contained simulation run}
\usage{
crun_sim_(
  nsteps = 500,
  progress_bar = FALSE,
  M = 120,
  env_mean_mc = 0,
  env_sd_mc = 1,
  cauchy = 1,
  trait_sds = 0,
  max_abun = 1e+06,
  npatch = 10,
  res = 1000,
  gradient = "random",
  K = 3,
  env_mean_lh = 0,
  env_sd_lh = 25,
  rho = 0.75,
  nind0 = 1000,
  recruitment = 0.05,
  dispersal = 0.05,
  reproduction = 0.1,
  expenditure = 0.1,
  res_input = 10,
  intake = 1,
  clustering = 1,
  dispersion = 1
)
}
\arguments{
\item{nsteps}{Positive integer, number of simulation time steps.}

\item{progress_bar}{Logical, whether to show the progress bar for the simulation run; default to FALSE.}

\item{M}{Integer, species richness within metacommunity, positive integer.}

\item{env_mean_mc}{Numeric, mean environmental factor value within the extent of a metacommunity.}

\item{env_sd_mc}{Non-negative numeric, variation of environmental factor value within the extent of a metacommunity.}

\item{cauchy}{Non-negative numeric, delta parameter in Cauchy function used to connect species' trait values and log-abundances.}

\item{trait_sds}{Non-negative numeric, intraspecific trait variation; either a single value if same for all species, or a M-length vector.}

\item{max_abun}{Numeric, the maximum abundance allowed for the most abundant species in the metacommunity.}

\item{npatch}{Positive integer, number of patches in the environment.}

\item{res}{Non-negative numeric, available resource level within each patch of the local habitat.}

\item{gradient}{Character, the rule by which patches get their values of env, either:
\code{gradient = "random"} - default value, env is an independent random variable drawn from N(env_mean, env_sd),
\code{gradient = "linear"} - env changes linearly from patch number 1 to patch number \code{npatch} with min and max drawn from 95% bound of N(env_mean, env_sd),
\code{gradient = "correlated"} - environmental factor is correlated with patch number with correlation coefficient equal \code{rho}, or
\code{gradient = "clustered"} - for K clusters, there are linearly distributed local env_means and small env_sd.}

\item{K}{Positive integer, number of clusters if \code{gradient = "clustered"}.}

\item{env_mean_lh}{Numeric, mean value. of the environmental factor across all habitat patches.}

\item{env_sd_lh}{Positive numeric, variation of the environmental factor across all habitat patches.}

\item{rho}{Positive numeric between \code{0} and \code{1}, correlation coefficient between patch number and environmental factor when \code{gradient = "correlated"}.}

\item{nind0}{Positive integer, community size at simulation initialization.}

\item{recruitment}{Non-negative double, recruitment rate (i.e., expectation of number of individuals recruited into local habitat per patch per time step). Feeds into \code{recruit()}.}

\item{dispersal}{Non-negative numeric, dispersal rate per habitat patch, i.e., expectation of the number of individuals per patch that disperse to a neighboring patch. Feeds into \code{dispersal()}.}

\item{reproduction}{Non-negative numeric; mass at which individuals reproduce. Feeds into \code{dem()}.}

\item{expenditure}{Numeric; amount of body mass each individual needs to spend in order to survive one time step. Feeds into \code{dem()}.}

\item{res_input}{Numeric, increment of the resource level within a time step. Feeds into \code{lh_input_res()}.}

\item{intake}{Positive numeric, single-meal resource intake by an individual in an event of a successful feeding. Feeds into \code{forage()}.}

\item{clustering}{Numeric, effect of niche clustering on probability of competition (between 0 and 1). Default to one.
If zero, local environmental conditions do not affect individual resource intake. Feeds into \code{forage()}.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment (between 0 and 1). Default to one.
If zero, trait differences among individuals do not affect resource intake. Feeds into \code{forage()}.}
}
\value{
Simulation output from \code{run_sim_()}.
}
\description{
Contained simulation run
}
