% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{dem}
\alias{dem}
\title{Simulate demographic processes within the local community}
\usage{
dem(com, metacom, age_crit = 10, mass_crit = 5)
}
\arguments{
\item{com}{A tibble (dataframe) representing the \strong{local} community with the following columns:
\code{patch} - identifier of habitat patch;
\code{species} - species ID;
\code{trait} - individual trait value;
\code{age} - current age of the individual;
\code{mass} - current body mass of the individual;
\code{lifespan} - maximum lifespan after reaching which the individual will perish;
\code{repmass} - body mass after reaching which the individual will asexually reproduce into two children individuals.}

\item{metacom}{A tibble (dataframe) representing the \strong{metacommunity} with the following columns:
\code{species} - species ID, in ascending order relative to the trait value,
\code{trait} - species mean trait value,
\code{abundance} - expected species abundance,
\code{trait_sd} - intraspecific trait variation.}

\item{age_crit}{Numeric; critical age at which half of the individuals die.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce.}
}
\value{
A tibble (dataframe) representing the **local community** with the following columns:
\code{patch} - identifier of habitat patch;
\code{species} - species ID;
\code{trait} - individual trait value;
\code{age} - current age of the individual;
\code{mass} - current body mass of the individual;
\code{lifespan} - maximum lifespan after reaching which the individual will perish;
\code{repmass} - body mass after reaching which the individual will asexually reproduce into two children individuals.
}
\description{
Take in the tibble representing the local metacommunity and return what it would look like after two demographic processes are applied:
(1) mortality, i.e., elimination of all individuals that exceed their maximum lifespan, and
(2) reproduction, at which all individuals whose body mass exceeds their reproductive body mass asexually reproduce into two new individuals
with body masses equal to half of that in parent individual and trait values equal to parent's trait value plus random error,
as dictated by the metacommunity description.

Note that age of individuals is not increased by this function, this is a responsibility of the \code{time_step()}.
}
\examples{
x = init_meta()
y = dplyr::mutate(draw_lcom(x, 100), patch = 1)
dem(com = y, metacom = x)

}
