% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_disperse.R
\name{disperse}
\alias{disperse}
\title{Simulate patch-to-patch dispersal}
\usage{
disperse(lc, lh, dispersal = 1)
}
\arguments{
\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{dispersal}{Non-negative numeric, dispersal rate per habitat patch, i.e., expectation of the number of individuals per patch that disperse to a neighboring patch.}
}
\value{
A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).
}
\description{
As the local habitat consists of a finite set of discrete patches, one might allow individuals to disperse across patches.
This function shuffles patch IDs in the input local community \code{"FilterABM_lc"} object in such a way that it simulates expected
\code{ndisperse} individuals per patch to move into a random neighboring patch.
Note that the exact number of individuals that "choose" to disperse is drawn from a Poisson distribution, where \code{ndisperse} serves as a parameter.
}
\examples{
mc1 = init_meta()
lh1 = init_envt()
lc1 = draw_lcom(mc = mc1, lh = lh1)
lc1 = recruit(lc = lc1, mc = mc1, lh = lh1)
lc1 = dem(lc = lc1, mc = mc1)
disperse(lc = lc1, lh = lh1, dispersal = 10)
}
