% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_lcom.R
\name{draw_lcom}
\alias{draw_lcom}
\title{Draw individual(s) from a metacommunity into local community}
\usage{
draw_lcom(metacom, nind, age_crit = 10, mass_crit = 5)
}
\arguments{
\item{metacom}{A tibble (dataframe) representing the \strong{metacommunity} with the following columns:
\code{species} - species ID, in ascending order relative to the trait value,
\code{trait} - species mean trait value,
\code{abundance} - expected species abundance,
\code{trait_sd} - intraspecific trait variation}

\item{nind}{A number of individuals to draw, positive integer}

\item{age_crit}{Numeric; critical age at which half of the individuals die}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce}
}
\value{
A tibble (dataframe) representing \code{nind} drawn individuals with the following columns:
\code{species} - species ID; \code{trait} - individual trait value; \code{age} - age of the individual; \code{mass} - body mass;
\code{lifespan} - maximum lifespan after reaching which the individual will perish;
\code{repmass} - body mass after reaching which the individual will asexually reproduce into two children individuals
}
\description{
Provided a metacommunity object (structured as generated by \code{init_meta()} function),
returns a tibble with the number of rows representing recruited individuals.
Rows in the output tibble represent indiviuals recruited from the metacommunity (nrows equal to \code{nind}).
Columns represent the species ID (arbitrary identifier), the trait value, age of the individual, its body mass, maximum lifespan, and reproductive body mass.

Unlike \code{recruit()}, this function will generate a tibble in a format almost ready to \code{bind_rows} with a tibble representing the local community.
}
\examples{
x = init_meta()
draw_lcom(x, nind = 1)

}
\seealso{
\code{init_meta()}, \code{recruit()}
}
