% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{fd95}
\alias{fd95}
\title{Helper, 95-percentile range of a values}
\usage{
fd95(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Range between the 2.5th and 9.75th percentile
}
\description{
Helper, 95-percentile range of a values
}
