% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_lh.R
\name{FilterABM_lh}
\alias{FilterABM_lh}
\title{Build a local habitat object}
\usage{
FilterABM_lh(x, val_in = TRUE, val_out = TRUE, gradient = NULL)
}
\arguments{
\item{x}{Optional (i.e., if not supplied, will return an empty valid object) tibble or dataframe with the columns specified above
(unique integer \code{patch}, double \code{env}).
Each row represents a discrete habitat patch within the local habitat with its own level of the environmental factor.}

\item{val_in}{Logical, default to TRUE. Is formal validation of \strong{input} necessary to check if consistent with the \code{"FilterABM_lh"} class?}

\item{val_out}{Logical, default to TRUE. Is formal validation of \strong{output} necessary to check if consistent with the \code{"FilterABM_lh"} class?}

\item{gradient}{Can be NULL. Character, the rule by which patches get their values of env, either:
\code{gradient = "random"} - default value, env is an independent random variable drawn from N(env_mean, env_sd),
\code{gradient = "linear"} - env changes linearly from patch number 1 to patch number \code{npatch} with min and max drawn from 95% bound of N(env_mean, env_sd),
\code{gradient = "correlated"} - environmental factor is correlated with patch number with correlation coefficient equal \code{rho}, or
\code{gradient = "clustered"} - for K clusters, there are linearly distributed local env_means and small env_sd.}
}
\value{
An object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Build a local habitat (\code{"FilterABM_lh"}) object.

The \code{"FilterABM_lh"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The local habitat object must have the following columns:
\code{patch} - unique integer, patch ID,
\code{env} - double, patch-specific level of the environmental factor.
}
\examples{
x = dplyr::tibble(patch = 1:10, env = rnorm(10))
FilterABM_lh(x)

}
