% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_mc.R
\name{FilterABM_mc}
\alias{FilterABM_mc}
\title{Build a metacommunity object}
\usage{
FilterABM_mc(x, val_in = TRUE, val_out = TRUE)
}
\arguments{
\item{x}{Optional (i.e., if not supplied, will return an empty valid object) tibble or dataframe with the columns specified above
(unique integer \code{species}, double \code{trait}, integer positive \code{abundance}, and non-negative double \code{trait_sd}).
Each row represents a species in the regional species pool.}

\item{val_in}{Logical, default to TRUE. Is formal validation of \strong{input} necessary to check if consistent with the \code{"FilterABM_mc"} class?}

\item{val_out}{Logical, default to TRUE. Is formal validation of \strong{output} necessary to check if consistent with the \code{"FilterABM_mc"} class?}
}
\value{
A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Build a metacommunity (\code{"FilterABM_mc"}) object.

The \code{"FilterABM_mc"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The metacommunity object must have the following columns:
\code{species} - integer, species ID, in ascending order relative to the trait value,
\code{trait} - double, species mean trait value,
\code{abundance} - integer, expected species abundance,
\code{trait_sd} - non-negative double, intraspecific trait variation.
}
\examples{
x = dplyr::tibble(species = 1:10, trait = rnorm(10), abundance = rpois(10, 10), trait_sd = 0)
FilterABM_mc(x)

}
