% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage.R
\name{forage}
\alias{forage}
\title{Simulate foraging within a habitat patch}
\usage{
forage(com, env, res, res_input, R, clustering = 1, dispersion = 1)
}
\arguments{
\item{com}{A tibble (dataframe) representing the \strong{local} community with the following columns:
\code{species} - species ID;
\code{trait} - individual trait value;
\code{age} - current age of the individual;
\code{mass} - current body mass of the individual;
\code{lifespan} - maximum lifespan after reaching which the individual will perish;
\code{repmass} - body mass after reaching which the individual will asexually reproduce into two children individuals.}

\item{env}{Numeric, value of the environmental factor in the patch.}

\item{res}{Numeric, value of the resource available within the patch.}

\item{res_input}{Numeric, increment of the resource level within a time step.}

\item{R}{Numeric, resource level at which all individuals within a community successfully consume the resource (i.e., probability of resource consumption equals one).}

\item{clustering}{Numeric, effect of niche clustering on probability of competition. Default to one.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment. Default to one.}
}
\value{
A list with the following elements:
\code{[["com"]]} - updated community object,
\code{[["res"]]} - available resource level,
\code{[["res_input"]]} - increment of the resource level within a time step,
\code{[["p_compete"]]} - sum of foraging probabilities given competition,
\code{[["p_consume"]]} - sum of foraging probabilities given resource availability.
}
\description{
The function takes in the description of the local community and surrounding environmental factors.
Based on these, each individual within a community takes a turn in random sequence consuming the available resource.
All else being equal, an individual with a trait value optimal for the local environmental factor \code{env} will consume more resource.
If there are many individuals in the community with a similar trait value, the focal individual is less likely to consume the resource.
Similarly, in an individuals-rich environment, the chances of consuming anything at all are reduced.
}
\examples{
x = init_meta()
y = draw_lcom(x, 100)
forage(com = y, env = 0, res = 1000, res_input = 10, R = 1000)

}
