% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_meta.R
\name{init_meta}
\alias{init_meta}
\title{Initialize a metacommunity}
\usage{
init_meta(M = 120, env_mean = 0, env_sd = 25, cauchy = 1.25, trait_sds = 0)
}
\arguments{
\item{M}{Species richness within metacommunity, positive integer}

\item{env_mean}{Mean environmental factor value within the extent of a metacommunity}

\item{env_sd}{Variation of environmental factor value within the extent of a metacommunity}

\item{cauchy}{Delta parameter in Cauchy function used to connect species' trait values and abundances}

\item{trait_sds}{Intraspecific trait variation; either a single value if same for all species, or a M-length vector}
}
\value{
A tibble (dataframe) with the following columns:
\code{species} - species ID, in ascending order relative to the trait value,
\code{trait} - species mean trait value,
\code{abundance} - expected species abundance.
}
\description{
Create a metacommunity "object", a tibble with rows describing different abstract hypothetical species in the species pool,
and columns representing species identifier, mean trait value across individuals within the species, and abundance of the species.
Intraspecific trait variation is set manually and passed to the output as a function parameter.
}
\examples{
init_meta()
}
