% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{jaccard}
\alias{jaccard}
\title{Helper, Jaccard index}
\usage{
jaccard(spp1, spp2)
}
\arguments{
\item{spp1}{Species identities in community 1}

\item{spp2}{Species identities in community 2}
}
\value{
Jaccard index
}
\description{
Helper, Jaccard index
}
