% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{jaccard_boot}
\alias{jaccard_boot}
\title{Helper, permutational expectation of Jaccard index for observed community and random draws from regional pool}
\usage{
jaccard_boot(obs, mc, nperm = 100)
}
\arguments{
\item{obs}{Observed community}

\item{mc}{A regional pool object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{nperm}{Numeric, number of permutations.}
}
\value{
Jaccard indices
}
\description{
Helper, permutational expectation of Jaccard index for observed community and random draws from regional pool
}
