% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_forage.R
\name{lh_input_res}
\alias{lh_input_res}
\title{Regenerate resource in local habitat patches}
\usage{
lh_input_res(lh, res_input = 0)
}
\arguments{
\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{res_input}{Numeric, increment of the resource level within a time step.}
}
\value{
A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).
}
\description{
Regenerate resource in local habitat patches
}
\examples{
lh1 = init_envt()
lh_input_res(lh = lh1, res_input = 10)

}
