% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_lc.R
\name{plot.FilterABM_lc}
\alias{plot.FilterABM_lc}
\title{Plot a local community object}
\usage{
\method{plot}{FilterABM_lc}(x, y, ...)
}
\arguments{
\item{x}{A valid local habitat (\code{"FilterABM_lc"}) object.}

\item{y}{Parameter two.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Plot a local community object
}
