% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FilterABM_lh.R
\name{plot.FilterABM_lh}
\alias{plot.FilterABM_lh}
\title{Plot a local habitat object}
\usage{
\method{plot}{FilterABM_lh}(x, y, ...)
}
\arguments{
\item{x}{A valid local habitat (\code{"FilterABM_lh"}) object.}

\item{y}{Parameter two.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Plot a local habitat object
}
