% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_mc.R
\name{plot.FilterABM_mc}
\alias{plot.FilterABM_mc}
\title{Plot a metacommunity object}
\usage{
\method{plot}{FilterABM_mc}(x, y, ...)
}
\arguments{
\item{x}{A valid metacommunity (\code{"FilterABM_mc"}) object.}

\item{y}{Parameter two.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Plot a metacommunity object
}
