% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{plot.FilterABM_runsim}
\alias{plot.FilterABM_runsim}
\title{Diagnostic plots for a simulation output (for long output)}
\usage{
\method{plot}{FilterABM_runsim}(x, y, ...)
}
\arguments{
\item{x}{An output of \code{run_sim_()}.}

\item{y}{Parameter two.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Diagnostic plots for a simulation output (for long output)
}
