% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{plot_run_sim}
\alias{plot_run_sim}
\title{Plot a simulation output}
\usage{
plot_run_sim(runsim)
}
\arguments{
\item{runsim}{An output of \code{run_sim()}.}
}
\value{
A plot of simulation parameters with time steps.
}
\description{
Plot a simulation output
}
