% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{plot_run_sim_}
\alias{plot_run_sim_}
\title{Plot a simulation output (for long output)}
\usage{
plot_run_sim_(runsim)
}
\arguments{
\item{runsim}{An output of \code{run_sim_()}.}
}
\value{
A plot of simulation parameters with time steps.
}
\description{
Plot a simulation output (for long output)
}
