% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_recruit.R
\name{pred_logit}
\alias{pred_logit}
\title{Helper function, logit}
\usage{
pred_logit(p, crit)
}
\arguments{
\item{p}{A random value between 0 and 1}

\item{crit}{A critical value corresponding to p = 0.5}
}
\value{
A random value following the logit function
}
\description{
Logit-transformation of an input variable.
}
\examples{
pred_logit(runif(1), crit = 0.5)
}
