% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{recruit}
\alias{recruit}
\title{Add new individuals from the metacommunity into the local community}
\usage{
recruit(lc, mc, lh, nind = 1, ...)
}
\arguments{
\item{lc}{An object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" with the following columns:
\code{species} - integer for species ID,
\code{trait} - double for individual trait value,
\code{age} - integer for the current individual age in time steps,
\code{mass} - double for the current individual body mass,
\code{lifespan} - double for maximum individual age,
\code{repmass} - double for a critical body mass at which the individual reproduces,
\code{patch} - integer for the patch ID in which individual currently resides.}

\item{mc}{An object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" with the following columns:
\code{species} - integer, species ID, in ascending order relative to the trait value,
\code{trait} - double, species mean trait value,
\code{abundance} - integer, expected species abundance,
\code{trait_sd} - non-negative double, intraspecific trait variation.}

\item{lh}{An object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" with the following columns:
\code{patch} - unique integer, patch ID,
\code{env} - double, patch-specific level of the environmental factor.}

\item{nind}{A number of individuals to draw, positive integer.}

\item{...}{Other parameters applicable to \code{draw_lcom()}.}
}
\value{
An object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Provided a metacommunity object (class \code{"FilterABM_mc")}) and the local habitat (\code{"FilterABM_lh"}),
as well as a prior local community (\code{"FilterABM_lc"}),
add the newly drawn individuals into the metacommunity.
Rows in the output tibble represent individuals recruited from the metacommunity (nrows equal to \code{nind}).
If the species is described as one having no intraspecific trait variation (e.g., \code{trait_sd = 0}),
then the trait of a recruited individual will be equal to the species-specific trait value (e.g., \code{trait}), otherwise
a small random error will be added drawn from a normal distribution with SD equal to \code{trait_sd}.
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh)
recruit(lc = lc, mc = mc, lh = lh)

}
