% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{rkde}
\alias{rkde}
\title{Helper, draw random points from a kernel density estimation}
\usage{
rkde(kde, n = 1)
}
\arguments{
\item{kde}{A kernel density estimation as generated by density()}

\item{n}{Number of values to draw}
}
\value{
Random values drawn from an empirical distribution
}
\description{
Helper, draw random points from a kernel density estimation
}
