% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FilterABM.R
\name{run_FilterABM}
\alias{run_FilterABM}
\title{Run the simulation with enclosed evaluation}
\usage{
run_FilterABM(
  nsteps,
  M = 120,
  env_mean_mc = 0,
  env_sd_mc = 25,
  cauchy = 1.5,
  trait_sds = 1,
  max_abun = 1e+06,
  npatch = 100,
  res = 1000,
  gradient = "correlated",
  K = 3,
  env_mean_lh = 10,
  env_sd_lh = 1,
  rho = 0.75,
  recr_init = 100,
  age_crit = 10,
  mass_crit = 5,
  recruitment = 0.05,
  dispersal = 0.05,
  res_input = 10,
  R = 1000,
  clustering = 1,
  dispersion = 1
)
}
\arguments{
\item{nsteps}{Positive integer, number of simulation time steps.}

\item{M}{Integer, species richness within metacommunity, positive integer. Feeds into \code{init_meta()}.}

\item{env_mean_mc}{Numeric, mean environmental factor value within the extent of a metacommunity. Feeds into \code{init_meta()}.}

\item{env_sd_mc}{Non-negative numeric, variation of environmental factor value within the extent of a metacommunity. Feeds into \code{init_meta()}.}

\item{cauchy}{Non-negative numeric, delta parameter in Cauchy function used to connect species' trait values and log-abundances. Feeds into \code{init_meta()}.}

\item{trait_sds}{Non-negative numeric, intraspecific trait variation; either a single value if same for all species, or a M-length vector. Feeds into \code{init_meta()}.}

\item{max_abun}{Numeric, the maximum abundance allowed for the most abundant species in the metacommunity. Feeds into \code{init_meta()}.}

\item{npatch}{Positive integer, number of patches in the environment. Feeds into \code{init_envt()}.}

\item{res}{Non-negative numeric, available resource level within each patch of the local habitat. Feeds into \code{init_envt()}.}

\item{gradient}{Feeds into \code{init_envt()}. Character, the rule by which patches get their values of env, either:
\code{gradient = "random"} - default value, env is an independent random variable drawn from N(env_mean, env_sd),
\code{gradient = "linear"} - env changes linearly from patch number 1 to patch number \code{npatch} with min and max drawn from 95% bound of N(env_mean, env_sd),
\code{gradient = "correlated"} - environmental factor is correlated with patch number with correlation coefficient equal \code{rho}, or
\code{gradient = "clustered"} - for K clusters, there are linearly distributed local env_means and small env_sd.}

\item{K}{Positive integer, number of clusters if \code{gradient = "clustered"}. Feeds into \code{init_envt()}.}

\item{env_mean_lh}{Numeric, mean value. of the environmental factor across all habitat patches. Feeds into \code{init_envt()}.}

\item{env_sd_lh}{Positive numeric, variation of the environmental factor across all habitat patches. Feeds into \code{init_envt()}.}

\item{rho}{Positive numeric between \code{0} and \code{1}, correlation coefficient between patch number and environmental factor when \code{gradient = "correlated"}. Feeds into \code{init_envt()}.}

\item{recr_init}{Positive integer, a number of individuals to draw from the metacommunity. Feeds into \code{draw_lcom()}.}

\item{age_crit}{Numeric; critical age at which half of the individuals die. Feeds into \code{draw_lcom()} and \code{dem()}.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce. Feeds into \code{draw_lcom()} and \code{dem()}.}

\item{recruitment}{Non-negative double, recruitment rate (i.e., expectation of number of individuals recruited into local habitat per patch per time step). Feeds into \code{recruit()}.}

\item{dispersal}{Non-negative numeric, dispersal rate per habitat patch, i.e., expectation of the number of individuals per patch that disperse to a neighboring patch. Feeds into \code{dispersal()}.}

\item{res_input}{Numeric, increment of the resource level within a time step. Feeds into \code{lh_input_res()}.}

\item{R}{Numeric, resource level at which all individuals within a community successfully consume the resource (i.e., probability of resource consumption equals one). Feeds into \code{forage()}.}

\item{clustering}{Numeric, effect of niche clustering on probability of competition (between 0 and 1). Default to one. Feeds into \code{forage()}.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment (between 0 and 1). Default to one. Feeds into \code{forage()}.}
}
\value{
A list.
}
\description{
A slightly overparametrized function.
The objective is to make a model run extremely user-friendly, so that only the simulation parameters are supplied, and all objects needed for the simulation are evaluated internally.

Usage is not recommended, please use \code{run_sim()} instead.
}
