% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{[Long output] Run a simulation of environmental filtering on specified metacommunity, local habitat, and local community objects}
\usage{
run_sim(
  mc,
  lh,
  lc,
  nsteps = 500,
  progress_bar = FALSE,
  recruitment = 0.05,
  dispersal = 0.05,
  reproduction = 0.1,
  expenditure = 0.1,
  res_input = 10,
  intake = 1,
  clustering = 1,
  dispersion = 1
)
}
\arguments{
\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{nsteps}{Positive integer, number of simulation time steps.}

\item{progress_bar}{Logical, whether to show the progress bar for the simulation run; default to FALSE.}

\item{recruitment}{Non-negative double, recruitment rate (i.e., expectation of number of individuals recruited into local habitat per patch per time step). Feeds into \code{recruit()}.}

\item{dispersal}{Non-negative numeric, dispersal rate per habitat patch, i.e., expectation of the number of individuals per patch that disperse to a neighboring patch. Feeds into \code{dispersal()}.}

\item{reproduction}{Non-negative numeric; mass at which individuals reproduce. Feeds into \code{dem()}.}

\item{expenditure}{Numeric; amount of body mass each individual needs to spend in order to survive one time step. Feeds into \code{dem()}.}

\item{res_input}{Numeric, increment of the resource level within a time step. Feeds into \code{lh_input_res()}.}

\item{intake}{Positive numeric, single-meal resource intake by an individual in an event of a successful feeding. Feeds into \code{forage()}.}

\item{clustering}{Numeric, effect of niche clustering on probability of competition (between 0 and 1). Default to one.
If zero, local environmental conditions do not affect individual resource intake. Feeds into \code{forage()}.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment (between 0 and 1). Default to one.
If zero, trait differences among individuals do not affect resource intake. Feeds into \code{forage()}.}
}
\value{
A list with a quasi-"FilterABM_lc" local community object that has an additional column for time step ([["lcs"]]).
Additionally, habitat history at each time step ([["habitat"]]), and the total resource consumed per time step ([["consumed_resource"]]).
Input arguments included in the output as well.
}
\description{
Requires an input of pre-defined metacommunity, local habitat, and local community objects, the two latter of which are iteratively modified to reflect changes caused by environmental filtering.
This function is not fully enclosed, but expects pre-initialized objects.
This allows running multiple simulations on identical input.

Unlike \code{run_sim_()}, it saves the data from each time step, which does provide more insight on trait distribution across individuals with time, but is more memory-intensive.
}
