% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{sad_ks}
\alias{sad_ks}
\title{Helper, Kolmogorov-Smirnov statistic for two empirical species abundance distributions}
\usage{
sad_ks(sad1, sad2)
}
\arguments{
\item{sad1}{Vector, abundances of species in one community}

\item{sad2}{Vector, abundances of species in the second community}
}
\value{
Kolmogorov-Smirnov statistic
}
\description{
Helper, Kolmogorov-Smirnov statistic for two empirical species abundance distributions
}
