% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_init_envt.R
\name{simcor}
\alias{simcor}
\title{Helper, generate a random variable with set correlation coefficient to a focal variable}
\usage{
simcor(x, ymean = 0, ysd = 1, correlation = 0)
}
\arguments{
\item{x}{Numeric, input vector.}

\item{ymean}{Numeric, mean of the output vector.}

\item{ysd}{Positive numeric, SD of the output vector.}

\item{correlation}{Correlation coefficient.}
}
\value{
Numeric of the same length as \code{x}.
}
\description{
For a given vector \code{x}, generate an independent random vector of the same length such that correlation between \code{x} and output vector equals to a set coefficient \code{correlation}.
}
\examples{
x = rnorm(100)
simcor(x)

}
