% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_lc.R
\name{summary.FilterABM_lc}
\alias{summary.FilterABM_lc}
\title{Summarize a local community object}
\usage{
\method{summary}{FilterABM_lc}(object, ...)
}
\arguments{
\item{object}{A valid local community (\code{"FilterABM_lc"}) object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Summarize a local community object
}
