% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_FilterABM_lh.R
\name{summary.FilterABM_lh}
\alias{summary.FilterABM_lh}
\title{Summarize a local habitat object}
\usage{
\method{summary}{FilterABM_lh}(object, ...)
}
\arguments{
\item{object}{A valid local habitat (\code{"FilterABM_lh"}) object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Summarize a local habitat object
}
