% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_FilterABM_mc.R
\name{summary.FilterABM_mc}
\alias{summary.FilterABM_mc}
\title{Summarize a metacommunity object}
\usage{
\method{summary}{FilterABM_mc}(object, ...)
}
\arguments{
\item{object}{A valid metacommunity (\code{"FilterABM_mc"}) object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Summarize a metacommunity object
}
