% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{tad_ks}
\alias{tad_ks}
\title{Helper, Kolmogorov-Smirnov statistic for two empirical KDE distributions}
\usage{
tad_ks(tad1, tad2)
}
\arguments{
\item{tad1}{Kernel density estimated pdf of trait values in a community}

\item{tad2}{Kernel density estimated pdf (note: from and to arguments in stat::density() should be the same as for tad1)}
}
\value{
Kolmogorov-Smirnov statistic
}
\description{
Helper, Kolmogorov-Smirnov statistic for two empirical KDE distributions
}
