% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{tad_ks_boot}
\alias{tad_ks_boot}
\title{Helper, permutational expectation of Kolmogorov-Smirnov statistic for an observed TAD and TAD from random sample of regional pool}
\usage{
tad_ks_boot(tad, mc, nperm = 100)
}
\arguments{
\item{tad}{Kernel density estimated pdf (output of stats::density() function)}

\item{mc}{A regional pool object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{nperm}{Numeric, number of permutations.}
}
\value{
Kolmogorov-Smirnov statistic for each permutation
}
\description{
Helper, permutational expectation of Kolmogorov-Smirnov statistic for an observed TAD and TAD from random sample of regional pool
}
