% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_step.R
\name{time_step}
\alias{time_step}
\title{Simulate a discrete time step}
\usage{
time_step(
  com,
  metacom,
  envt,
  age_crit = 10,
  mass_crit = 5,
  res = 1000,
  res_input = 10,
  R = 1000,
  recruitment = 1,
  dispersal = 1
)
}
\arguments{
\item{com}{A tibble (dataframe) representing the \strong{local} community with the following columns:
\code{species} - species ID;
\code{trait} - individual trait value;
\code{age} - current age of the individual;
\code{mass} - current body mass of the individual;
\code{lifespan} - maximum lifespan after reaching which the individual will perish;
\code{repmass} - body mass after reaching which the individual will asexually reproduce into two children individuals.}

\item{metacom}{A tibble (dataframe) with the following columns:
\code{species} - species ID, in ascending order relative to the trait value,
\code{trait} - species mean trait value,
\code{abundance} - expected species abundance.
Output of \code{init_meta()}.}

\item{envt}{A tibble (dataframe) representing habitat patches (column \code{patch}) and the patch-specific value of the environmental factor (column \code{env}).
Output of \code{init_envt()}.}

\item{age_crit}{Numeric; critical age at which half of the individuals die.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce.}

\item{res}{available resource level, a dataframe with two columns:}

\item{res_input}{Numeric, increment of the resource level within a time step.}

\item{R}{Numeric, resource level at which all individuals within a community successfully consume the resource (i.e., probability of resource consumption equals one).}

\item{recruitment}{Numeric, expected number of individuals to be recruited from the metacommunity into one patch with \code{recruit()}.
Default to one.}

\item{dispersal}{Numeric, expected number of individuals to disperse into a different patch from their patch. Default to one.}
}
\value{
A list with the following elements:
\code{[[i]]} - habitat patch number with a list of the following:
\code{[[i]][["com"]]} - updated community object,
\code{[[i]][["res"]]} - updated resource value,
\code{[[i]][["res_consumed"]]} - sum of consumed resource,
\code{[[i]][["p_compete"]]} - sum of foraging probabilities given competition,
\code{[[i]][["p_consume"]]} - sum of foraging probabilities given resource availability.
}
\description{
Simulate changes in the local community through a single time step.

There are multiple processes involved in simulated changes in local community:
(1) agents are recruited from the metacommunity with \code{recruit()},
(2) age of all agents increases by one step and all agents whose age is greater than their maximum lifespan are eliminated with \code{dem()},
(3) all agents whose body mass is greater than individual reproductive body mass reproduce with \code{dem()},
(4) agents disperse to neighboring habitat patches at a set \code{dispersal} rate,
(5) agents consume resources with \code{forage()}.

An individual’s ability to consume the resource within a time step will depend on
(1) whether an individual’s trait value matches a cell’s environmental level as a proxy of how adapted to the local environment the individual is, and
(2) whether other individuals in the cell have similar trait values, thus inducing competition for the resource.

This function summarizes all actions commited by individuals into shaping the local community at a time step.
}
\examples{
x = init_meta()
y = draw_lcom(x, 100)
y$patch = sample(1:10, 100, replace = TRUE)
z = init_envt()
res0 <- dplyr::tibble(patch = 1:10, res = 1000)
time_step(com = y, metacom = x, envt = z, res = res0)

}
