% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterABM_lh.R
\name{validate_FilterABM_lh}
\alias{validate_FilterABM_lh}
\title{Local habitat validator}
\usage{
validate_FilterABM_lh(x)
}
\arguments{
\item{x}{A valid object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame"}
}
\value{
An object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Validate a local habitat (\code{"FilterABM_lh"}) object.

The \code{"FilterABM_lh"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The local habitat object must have the following columns:
\code{patch} - unique integer, patch ID,
\code{env} - double, patch-specific level of the environmental factor.
}
\examples{
x = dplyr::tibble(patch = 1:10, env = rnorm(10))
validate_FilterABM_lh(x)

}
