% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_FilterABM_mc.R
\name{validate_FilterABM_mc}
\alias{validate_FilterABM_mc}
\title{Metacommunity validator}
\usage{
validate_FilterABM_mc(x)
}
\arguments{
\item{x}{A valid metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame"}
}
\value{
A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Validate a metacommunity (\code{"FilterABM_mc"}) object.

The \code{"FilterABM_mc"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The metacommunity object must have the following columns:
\code{species} - integer, species ID, in ascending order relative to the trait value,
\code{trait} - double, species mean trait value,
\code{abundance} - integer, expected species abundance,
\code{trait_sd} - non-negative double, intraspecific trait variation.
}
\examples{
x = dplyr::tibble(species = 1:10, trait = rnorm(10), abundance = rpois(10, 10), trait_sd = 0)
validate_FilterABM_mc(x)

}
