% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_dem.R
\name{dem}
\alias{dem}
\title{Simulate demographic processes within the local community}
\usage{
dem(lc, mc, reproduction = 10, expenditure = 0.1)
}
\arguments{
\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{reproduction}{Non-negative numeric; mass at which individuals reproduce.}

\item{expenditure}{Positive numeric; amount of body mass each individual needs to spend in order to survive one time step.}
}
\value{
A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Take in the tibble representing the local metacommunity and return what it would look like after two demographic processes are applied:
(1) mortality, i.e., elimination of all individuals whose body mass is insufficient to survive, and
(2) reproduction, at which all individuals whose body mass exceeds their reproductive body mass asexually reproduce into two new individuals
with body masses equal to half of that in parent individual and trait values equal to parent's trait value plus random error,
as dictated by the metacommunity description.
}
\examples{
mc1 = init_meta()
lh1 = init_envt()
lc1 = draw_lcom(mc = mc1, lh = lh1)
lc1 = recruit(lc = lc1, mc = mc1, lh = lh1)
dem(lc = lc1, mc = mc1)

}
