% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.R
\name{dem}
\alias{dem}
\title{Simulate demographic processes within the local community}
\usage{
dem(lc, mc, age_crit = 10, mass_crit = 5)
}
\arguments{
\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{age_crit}{Numeric; critical age at which half of the individuals die.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce.}
}
\value{
A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).
}
\description{
Take in the tibble representing the local metacommunity and return what it would look like after two demographic processes are applied:
(1) mortality, i.e., elimination of all individuals that exceed their maximum lifespan, and
(2) reproduction, at which all individuals whose body mass exceeds their reproductive body mass asexually reproduce into two new individuals
with body masses equal to half of that in parent individual and trait values equal to parent's trait value plus random error,
as dictated by the metacommunity description.

Note that age of individuals is not increased by this function, this is a responsibility of the \code{adv_age()}.
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh)
lc = recruit(lc = lc, mc = mc, lh = lh)
lc = adv_age(lc = lc)
dem(lc = lc, mc = mc)

}
