% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{draw_lcom}
\alias{draw_lcom}
\title{Draw individual(s) from the metacommunity}
\usage{
draw_lcom(mc, lh, nind = 1, age_crit = 10, mass_crit = 5)
}
\arguments{
\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{nind}{Positive integer, a number of individuals to draw from the metacommunity.}

\item{age_crit}{Numeric; critical age at which half of the individuals die.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce.}
}
\value{
An object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).
}
\description{
Provided a metacommunity object (class \code{"FilterABM_mc")}) and the local habitat (\code{"FilterABM_lh"}),
returns a tibble with the number of rows representing recruited individuals.
Rows in the output tibble represent individuals recruited from the metacommunity (nrows equal to \code{nind}).
Columns represent the species ID (arbitrary identifier), the trait value, age of the individual, its body mass, maximum lifespan, and reproductive body mass,
as well as the local habitat patch occupied by an individual.

The major difference from \code{recruit()} is that \code{draw_meta()} returns a set of individuals drawn from the metacommunity,
while \code{recruit()} mutates the local community adding new individuals.
}
\examples{
mc = init_meta()
lh = init_envt()
draw_lcom(mc = mc, lh = lh)

}
