% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_FilterABM_lc.R
\name{FilterABM_lc}
\alias{FilterABM_lc}
\title{Build a local community object}
\usage{
FilterABM_lc(x, val_in = TRUE, val_out = TRUE)
}
\arguments{
\item{x}{Optional (i.e., if not supplied, will return an empty valid object) tibble or dataframe with the columns specified above
(integer \code{species}, double \code{trait}, double \code{mass}, integer \code{patch}).
Each row represents an individual in the local community.}

\item{val_in}{Logical, default to TRUE. Is formal validation of \strong{input} necessary to check if consistent with the \code{"FilterABM_lc"} class?}

\item{val_out}{Logical, default to TRUE. Is formal validation of \strong{output} necessary to check if consistent with the \code{"FilterABM_lc"} class?}
}
\value{
A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Create a local community (\code{"FilterABM_lc"}) object.
}
\details{
The \code{"FilterABM_lc"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The local community object must have the following columns:
\code{species} - integer; species ID,
\code{trait} - numeric; individual trait value,
\code{mass} - numeric; the current individual body mass,
\code{patch} - integer; the patch ID in which individual currently resides.
}
\examples{
FilterABM_lc()

}
