% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage.R
\name{forage}
\alias{forage}
\title{Simulate foraging within a habitat patch}
\usage{
forage(lc, lh, R, clustering = 1, dispersion = 1)
}
\arguments{
\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{R}{Numeric, resource level at which all individuals within a community successfully consume the resource (i.e., probability of resource consumption equals one).}

\item{clustering}{Numeric, effect of niche clustering on probability of competition. Default to one.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment. Default to one.}
}
\value{
A named list with:
\code{[["lc"]]} - A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc});
\code{[["lh"]]} - A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).
}
\description{
The function takes in the description of the local community and surrounding environmental factors.
Based on these, each individual within a community takes a turn in random sequence consuming the available resource.
All else being equal, an individual with a trait value optimal for the local environmental factor \code{env} will consume more resource.
If there are many individuals in the community with a similar trait value, the focal individual is less likely to consume the resource.
Similarly, in an individuals-rich environment, the chances of consuming anything at all are reduced.
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh, nind = 100)
lc = recruit(lc = lc, mc = mc, lh = lh, nind = 100)
lc = adv_age(lc = lc)
lc = dem(lc = lc, mc = mc)
lc = disperse(lc = lc, lh = lh, dispersal = 10)
forage(lc = lc, lh = lh, R = 1000)

}
