% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_envt.R
\name{init_envt}
\alias{init_envt}
\title{Initialize the local environment}
\usage{
init_envt(
  npatch = 10,
  gradient = "random",
  K = 3,
  env_mean = 0,
  env_sd = 25,
  rho = 0.75
)
}
\arguments{
\item{npatch}{Positive integer, number of patches in the environment.}

\item{gradient}{Character, the rule by which patches get their values of env, either:
\code{gradient = "random"} - default value, env is an independent random variable drawn from N(env_mean, env_sd),
\code{gradient = "linear"} - env changes linearly from patch number 1 to patch number \code{npatch} with min and max drawn from 95% bound of N(env_mean, env_sd),
\code{gradient = "correlated"} - environmental factor is correlated with patch number with correlation coefficient equal \code{rho}, or
\code{gradient = "clustered"} - for K clusters, there are linearly distributed local env_means and small env_sd.}

\item{K}{Positive integer, number of clusters if \code{gradient = "clustered"}.}

\item{env_mean}{Numeric, mean value. of the environmental factor across all habitat patches.}

\item{env_sd}{Positive numeric, variation of the environmental factor across all habitat patches.}

\item{rho}{Positive numeric between \code{0} and \code{1}, correlation coefficient between patch number and environmental factor when \code{gradient = "correlated"}.}
}
\value{
A tibble (dataframe) representing habitat patches (column \code{patch}) and the patch-specific value of the environmental factor (column \code{env}).
}
\description{
Create a tibble (dataframe) representing the local habitat as a set of habitat patches with slightly different levels of the environmental factor.
}
\examples{
init_envt()

}
