% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_meta.R
\name{init_meta}
\alias{init_meta}
\title{Initialize a metacommunity}
\usage{
init_meta(
  M = 120,
  env_mean_mc = 0,
  env_sd_mc = 1,
  cauchy = 1,
  trait_sds = 0,
  max_abun = 1e+06
)
}
\arguments{
\item{M}{Integer, species richness within metacommunity, positive integer.}

\item{env_mean_mc}{Numeric, mean environmental factor value within the extent of a metacommunity.}

\item{env_sd_mc}{Non-negative numeric, variation of environmental factor value within the extent of a metacommunity.}

\item{cauchy}{Non-negative numeric, delta parameter in Cauchy function used to connect species' trait values and log-abundances.}

\item{trait_sds}{Non-negative numeric, intraspecific trait variation; either a single value if same for all species, or a M-length vector.}

\item{max_abun}{Numeric, the maximum abundance allowed for the most abundant species in the metacommunity.}
}
\value{
A metacommunity \code{"FilterABM_mc"}-class tibble (see \code{?FilterABM::FilterABM_mc}).
}
\description{
Create a metacommunity (\code{"FilterABM_mc"}) object, a tibble with rows describing different abstract hypothetical species in the species pool,
and columns representing species identifier, mean trait value across individuals within the species, and abundance of the species.
Intraspecific trait variation is set manually and passed to the output as a function parameter.
}
\examples{
x = init_meta()
class(x)

}
