% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{recruit}
\alias{recruit}
\title{Draw a set number of individuals from the metacommunity}
\usage{
recruit(metacom, nind = 1)
}
\arguments{
\item{metacom}{A tibble (dataframe) representing the \strong{metacommunity} with the following columns:
\code{species} - species ID, in ascending order relative to the trait value,
\code{trait} - species mean trait value,
\code{abundance} - expected species abundance,
\code{trait_sd} - intraspecific trait variation.}

\item{nind}{A number of individuals to draw, positive integer.}
}
\value{
A tibble (dataframe) with the following columns: \code{species} - species ID, \code{trait} - individual trait value.
}
\description{
Provided a metacommunity object (structured as generated by \code{init_meta()} function),
returns a tibble with the number of rows representing recruited individuals.
Rows in the output tibble represent indiviuals recruited from the metacommunity (nrows equal to \code{nind}).
Columns represent the species ID (arbitrary identifier), and the trait value.
If the species is described as one having no intraspecific trait variation (e.g., \code{trait_sd = 0}),
then the trait of a recruited individual will be equal to the species-specific trait value (e.g., \code{trait}), otherwise
a small random error will be added drawn from a normal distribution with SD equal to \code{trait_sd}.
}
\examples{
x = init_meta()
recruit(x)

}
\seealso{
[init_meta()]
}
