% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit.R
\name{recruit}
\alias{recruit}
\title{Add new individuals from the metacommunity into the local community}
\usage{
recruit(lc, mc, lh, recruitment = 0, ...)
}
\arguments{
\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc})}

\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{recruitment}{Non-negative double, recruitment rate (i.e., expectation of number of individuals recruited into local habitat per patch per time step).}

\item{...}{Other parameters applicable to \code{draw_lcom()}.}
}
\value{
An object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame".
}
\description{
Provided a metacommunity object (class \code{"FilterABM_mc")}) and the local habitat (\code{"FilterABM_lh"}),
as well as a prior local community (\code{"FilterABM_lc"}),
add the newly drawn individuals into the metacommunity.
Rows in the output tibble represent individuals recruited from the metacommunity (nrows equal to \code{nind}).
If the species is described as one having no intraspecific trait variation (e.g., \code{trait_sd = 0}),
then the trait of a recruited individual will be equal to the species-specific trait value (e.g., \code{trait}), otherwise
a small random error will be added drawn from a normal distribution with SD equal to \code{trait_sd}.
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh)
recruit(lc = lc, mc = mc, lh = lh)

}
