% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation of environmental filtering on specified metacommunity, local habitat, and local community objects}
\usage{
run_sim(
  mc,
  lh,
  lc,
  nsteps = 500,
  progress_bar = FALSE,
  age_crit = 10,
  mass_crit = 5,
  recruitment = 0.05,
  dispersal = 0.05,
  res_input = 10,
  R = 1000,
  clustering = 1,
  dispersion = 1
)
}
\arguments{
\item{mc}{A metacommunity object of class "FilterABM_mc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_mc}).}

\item{lh}{A local habitat object of class "FilterABM_lh"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lh}).}

\item{lc}{A local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame" (see \code{?FilterABM::FilterABM_lc}).}

\item{nsteps}{Positive integer, number of simulation time steps.}

\item{progress_bar}{Logical, whether to show the progress bar for the simulation run; default to FALSE.}

\item{age_crit}{Numeric; critical age at which half of the individuals die. Feeds into \code{draw_lcom()} and \code{dem()}.}

\item{mass_crit}{Numeric; critical mass at which half of the individuals reproduce. Feeds into \code{draw_lcom()} and \code{dem()}.}

\item{recruitment}{Non-negative double, recruitment rate (i.e., expectation of number of individuals recruited into local habitat per patch per time step). Feeds into \code{recruit()}.}

\item{dispersal}{Non-negative numeric, dispersal rate per habitat patch, i.e., expectation of the number of individuals per patch that disperse to a neighboring patch. Feeds into \code{dispersal()}.}

\item{res_input}{Numeric, increment of the resource level within a time step. Feeds into \code{lh_input_res()}.}

\item{R}{Numeric, resource level at which all individuals within a community successfully consume the resource (i.e., probability of resource consumption equals one). Feeds into \code{forage()}.}

\item{clustering}{Numeric, effect of niche clustering on probability of competition (between 0 and 1). Default to one. Feeds into \code{forage()}.}

\item{dispersion}{Numeric, effect of niche dispersion on probability of trait filtering by the environment (between 0 and 1). Default to one. Feeds into \code{forage()}.}
}
\value{
A list of simulation parameters recorded at each step:
distribution of resource across habitat patches ([[habitat]]),
species area distribution of the local community ([[sad]]),
KDE trait distribution ([[traits]]),
and the total amount of resource consumed ([[consumed_resource]]).
}
\description{
Requires an input of pre-defined metacommunity, local habitat, and local community objects, the two latter of which are iteratively modified to reflect changes caused by environmental filtering.
Unlike \code{run_FilterABM()}, this function is not fully enclosed, but expects pre-initialized objects.
This allows running multiple simulations on identical input.
}
\examples{
mc = init_meta()
lh = init_envt()
lc = draw_lcom(mc = mc, lh = lh, nind = 100)
run_sim(mc, lh, lc, nsteps = 10)

}
