% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_FilterABM_lc.R
\name{validate_FilterABM_lc}
\alias{validate_FilterABM_lc}
\title{Local community validator}
\usage{
validate_FilterABM_lc(x)
}
\arguments{
\item{x}{A valid local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame"}
}
\value{
A valid local community object of class "FilterABM_lc"/"tbl_df"/"tbl"/"data.frame"
}
\description{
Validate a local community (\code{"FilterABM_lc"}) object.

The \code{"FilterABM_lc"} class is based on \code{dplyr}'s tibbles, thus \code{dplyr} functions are applicable.

The local community object must have the following columns:
\code{species} - integer; species ID,
\code{trait} - numeric; individual trait value,
\code{mass} - numeric; the current individual body mass,
\code{patch} - integer; the patch ID in which individual currently resides.
}
